% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\title{Run an R Makefile}
\usage{
make()
}
\description{
This function sources the \verb{_make.R} script located at the root of the project.
}
\examples{
make()
}
\author{
Nicolas Casajus, \email{nicolas.casajus@fondationbiodiversite.fr}
}
