% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-utils.R
\name{map_frame}
\alias{map_frame}
\title{Create a map frame}
\usage{
map_frame(xmin = -180, xmax = 180, ymin = -60, ymax = 90, crs = NA)
}
\arguments{
\item{xmin}{a numeric (min longitude of the frame)}

\item{xmax}{a numeric (max longitude of the frame)}

\item{ymin}{a numeric (min latitude of the frame)}

\item{ymax}{a numeric (max latitude of the frame)}

\item{crs}{a character of length 1 specifying the Coordinate Reference
System to project the frame. If \code{NA} (default) the frame is defined in
the WGS84 standard.}
}
\value{
An \code{sf} POLYGON.
}
\description{
Create a map frame
}
\examples{
robinson <- paste("+proj=robin +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84", 
                  "+datum=WGS84 +units=m +no_defs")
                  
land <- rnaturalearth::ne_countries(scale = "small", type = "countries", 
                                    returnclass = "sf")
                                    
land_rob <- sf::st_transform(land, robinson)

frame     <- map_frame(ymin = -90)
frame_rob <- map_frame(crs = robinson, ymin = -90)

par(mar = rep(0, 4), mfrow = c(2, 1))
plot(sf::st_geometry(frame))
plot(sf::st_geometry(land), add = TRUE)
plot(sf::st_geometry(frame_rob))
plot(sf::st_geometry(land_rob), add = TRUE)
}
