% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-utils.R
\name{map_graticules}
\alias{map_graticules}
\title{Create map graticules}
\usage{
map_graticules(
  xmin = -180,
  xmax = 180,
  ymin = -60,
  ymax = 90,
  meridians = seq(-160, 160, 40),
  parallels = seq(-60, 90, 20),
  crs = NA
)
}
\arguments{
\item{xmin}{a numeric (min longitude of the map extent)}

\item{xmax}{a numeric (max longitude of the map extent)}

\item{ymin}{a numeric (min latitude of the map extent)}

\item{ymax}{a numeric (max latitude of the map extent)}

\item{meridians}{a vector of longitudes at which meridians will be created}

\item{parallels}{a vector of latitudes at which parallels will be created}

\item{crs}{a character of length 1 specifying the Coordinate Reference
System to project graticules. If \code{NA} (default) graticules are defined in
the WGS84 standard.}
}
\value{
An \code{sf} MULTILINESTRING.
}
\description{
Create map graticules
}
\examples{
robinson <- paste("+proj=robin +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84", 
                  "+datum=WGS84 +units=m +no_defs")

frame <- map_frame(ymin = -90)
frame_rob <- map_frame(crs = robinson, ymin = -90)

grats     <- map_graticules(ymin = -90, parallels = seq( -80,  80, 20))
grats_rob <- map_graticules(crs = robinson, ymin = -90, 
                            parallels = seq( -80,  80, 20))

par(mar = rep(0, 4), mfrow = c(2, 1))
plot(sf::st_geometry(frame))
plot(sf::st_geometry(grats), add = TRUE)

plot(sf::st_geometry(frame_rob))
plot(sf::st_geometry(grats_rob), add = TRUE)
}
